sonar pass 17: drop redundant long cast, guard player location before dropItemNaturally

+++ mythicrod-paper/src/main/java/io/xcutiboo/mythicrod/paper/gui/menus/DropsMenu.java
@@ -616,7 +616,7 @@ public class DropsMenu extends BaseMenu {
             return 0;
         }
         long startIndex = (long) page * CONTENT_SLOTS.length;
-        long visible = Math.clamp(itemCount - startIndex, 0L, (long) CONTENT_SLOTS.length);
+        long visible = Math.clamp(itemCount - startIndex, 0L, CONTENT_SLOTS.length);
         return (int) visible;
     }
